major UI rework
