+++ app/src/main/java/dev/rodosteam/questtime/screen/common/editor/EditorFragment.kt
@@ -1,4 +1,4 @@
-package dev.rodosteam.questtime.ui.external
+package dev.rodosteam.questtime.screen.common.editor
 
 import android.os.Bundle
 import android.view.LayoutInflater
@@ -8,12 +8,13 @@ import android.widget.TextView
 import androidx.fragment.app.Fragment
 import androidx.lifecycle.Observer
 import androidx.lifecycle.ViewModelProvider
-import dev.rodosteam.questtime.databinding.FragmentExternalBinding
+import dev.rodosteam.questtime.databinding.FragmentEditorBinding
+import dev.rodosteam.questtime.screen.common.base.BaseFragment
 
-class ExternalFragment : Fragment() {
+class EditorFragment : BaseFragment() {
 
-    private lateinit var externalViewModel: ExternalViewModel
-    private var _binding: FragmentExternalBinding? = null
+    private lateinit var editorViewModel: EditorViewModel
+    private var _binding: FragmentEditorBinding? = null
 
     // This property is only valid between onCreateView and
     // onDestroyView.
@@ -24,14 +25,14 @@ class ExternalFragment : Fragment() {
         container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
-        externalViewModel =
-            ViewModelProvider(this).get(ExternalViewModel::class.java)
+        editorViewModel =
+            ViewModelProvider(this).get(EditorViewModel::class.java)
 
-        _binding = FragmentExternalBinding.inflate(inflater, container, false)
+        _binding = FragmentEditorBinding.inflate(inflater, container, false)
         val root: View = binding.root
 
-        val textView: TextView = binding.textExternal
-        externalViewModel.text.observe(viewLifecycleOwner, Observer {
+        val textView: TextView = binding.textEditor
+        editorViewModel.text.observe(viewLifecycleOwner, Observer {
             textView.text = it
         })
         return root
