+++ app/src/main/java/dev/rodosteam/questtime/screen/common/library/LibraryViewModel.kt
@@ -1,0 +1,10 @@
+package dev.rodosteam.questtime.screen.common.library
+
+import androidx.lifecycle.LiveData
+import androidx.lifecycle.MutableLiveData
+import androidx.lifecycle.ViewModel
+import dev.rodosteam.questtime.quest.model.QuestItem
+
+class LibraryViewModel() : ViewModel() {
+
+}
