--- app/src/main/java/dev/rodosteam/questtime/ui/settings/SettingsViewModel.kt
@@ -1,13 +1,0 @@
-package dev.rodosteam.questtime.ui.settings
-
-import androidx.lifecycle.LiveData
-import androidx.lifecycle.MutableLiveData
-import androidx.lifecycle.ViewModel
-
-class SettingsViewModel : ViewModel() {
-
-    private val _text = MutableLiveData<String>().apply {
-        value = "This is settings Fragment"
-    }
-    val text: LiveData<String> = _text
-}
