+++ app/src/main/res/layout/fragment_search_items_row.xml
